+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for PHP 7.1 (#967)
 
 ### Fixed
+- Allow `@import` after ignored invalid `@charset` (@1081)
 - Allow line feeds within `<html>` tag (#987)
 
 ## 5.0.1